+++ sdk/js/src/relayer/consts.ts
@@ -1,8 +1,8 @@
-import { ChainId, Network, ChainName, CHAIN_ID_TO_NAME } from "../";
 import { ethers } from "ethers";
+import { ChainName, Network } from "../";
 import {
-  WormholeRelayer__factory,
   WormholeRelayer,
+  WormholeRelayer__factory,
 } from "../ethers-contracts/";
 
 type AddressInfo = {
@@ -56,6 +56,26 @@ const TESTNET: { [K in ChainName]?: AddressInfo } = {
     wormholeRelayerAddress: "0xea8029CD7FCAEFFcD1F53686430Db0Fc8ed384E1",
     mockDeliveryProviderAddress: "0x60a86b97a7596eBFd25fb769053894ed0D9A8366",
     mockIntegrationAddress: "0x9Ee656203B0DC40cc1bA3f4738527779220e3998",
+  },
+  sepolia: {
+    wormholeRelayerAddress: "0x7B1bD7a6b4E61c2a123AC6BC2cbfC614437D0470",
+    mockDeliveryProviderAddress: "0x7A0a53847776f7e94Cc35742971aCb2217b0Db81",
+    mockIntegrationAddress: "0x68b7Cd0d27a6F04b2F65e11DD06182EFb255c9f0",
+  },
+  arbitrum_sepolia: {
+    wormholeRelayerAddress: "0x7B1bD7a6b4E61c2a123AC6BC2cbfC614437D0470",
+    mockDeliveryProviderAddress: "0x7A0a53847776f7e94Cc35742971aCb2217b0Db81",
+    mockIntegrationAddress: "0x2B1502Ffe717817A0A101a687286bE294fe495f7",
+  },
+  optimism_sepolia: {
+    wormholeRelayerAddress: "0x93BAD53DDfB6132b0aC8E37f6029163E63372cEE",
+    mockDeliveryProviderAddress: "0x7A0a53847776f7e94Cc35742971aCb2217b0Db81",
+    mockIntegrationAddress: "0xA404B69582bac287a7455FFf315938CCd92099c1",
+  },
+  base_sepolia: {
+    wormholeRelayerAddress: "0x93BAD53DDfB6132b0aC8E37f6029163E63372cEE",
+    mockDeliveryProviderAddress: "0x7A0a53847776f7e94Cc35742971aCb2217b0Db81",
+    mockIntegrationAddress: "0xA404B69582bac287a7455FFf315938CCd92099c1",
   },
 };
 
@@ -208,6 +228,10 @@ export const RPCS_BY_CHAIN: {
     gnosis: "https://sokol.poa.network/",
     rootstock: "https://public-node.rsk.co",
     base: "https://goerli.base.org",
+    sepolia: "https://rpc.ankr.com/eth_sepolia",
+    arbitrum_sepolia: "https://sepolia-rollup.arbitrum.io/rpc",
+    optimism_sepolia: "https://sepolia.optimism.io",
+    base_sepolia: "https://sepolia.base.org",
   },
   DEVNET: {
     ethereum: "http://localhost:8545",
@@ -243,12 +267,16 @@ export const getWormscanAPI = (_network: Network) => {
 };
 
 export const getNameFromCCTPDomain = (
-  domain: number
+  domain: number,
+  environment: Network = "MAINNET"
 ): ChainName | undefined => {
-  if (domain === 0) return "ethereum";
-  else if (domain === 1) return "avalanche";
-  else if (domain === 2) return "optimism";
-  else if (domain === 3) return "arbitrum";
-  else if (domain === 6) return "base";
+  if (domain === 0) return environment === "MAINNET" ? "ethereum" : "sepolia";
+  else if (domain === 1) "avalanche";
+  else if (domain === 2)
+    return environment === "MAINNET" ? "optimism" : "optimism_sepolia";
+  else if (domain === 3)
+    return environment === "MAINNET" ? "arbitrum" : "arbitrum_sepolia";
+  else if (domain === 6)
+    return environment === "MAINNET" ? "base" : "base_sepolia";
   else return undefined;
 };